feat: add measurement // measurements
Add a new grain that takes care of handling measurements. They are also
registered with the sample they belong to.

Note to self: F# data structures do not necessarily work with orleans
due to subtle ctor issues. F.ex. A map with a Guid key (which may be
a bad idea anyway) will fail to deserialize on grain activation due to
guid not having an ctor.